Modified data type in forecast time

+++ app/src/main/java/ca/nismit/util/weather/pojoForecast/List.java
@@ -8,7 +8,7 @@ public class List {
 
     @SerializedName("dt")
     @Expose
-    private Integer dt;
+    private Long dt;
     @SerializedName("main")
     @Expose
     private Main main;
@@ -34,11 +34,11 @@ public class List {
     @Expose
     private String dtTxt;
 
-    public Integer getDt() {
+    public Long getDt() {
         return dt;
     }
 
-    public void setDt(Integer dt) {
+    public void setDt(Long dt) {
         this.dt = dt;
     }
